+++ docs/project/changelog.rst
@@ -25,12 +25,52 @@ fixing regressions shortly after a release.
 Only documented APIs are public. Undocumented, private APIs may change without
 notice.
 
+.. _14.2:
+
+14.2
+----
+
+*In development*
+
+Bug fixes
+.........
+
+* Wrapped errors when reading the opening handshake request or response in
+  :exc:`~exceptions.InvalidMessage` so that :func:`~asyncio.client.connect`
+  raises :exc:`~exceptions.InvalidHandshake` or a subclass when the opening
+  handshake fails.
+
+.. _14.1:
+
+14.1
+----
+
+*November 13, 2024*
+
+Improvements
+............
+
+* Supported ``max_queue=None`` in the :mod:`asyncio` and :mod:`threading`
+  implementations for consistency with the legacy implementation, even though
+  this is never a good idea.
+
+* Added ``close_code`` and ``close_reason`` attributes in the :mod:`asyncio` and
+  :mod:`threading` implementations for consistency with the legacy
+  implementation.
+
+Bug fixes
+.........
+
+* Once the connection is closed, messages previously received and buffered can
+  be read in the :mod:`asyncio` and :mod:`threading` implementations, just like
+  in the legacy implementation.
+
 .. _14.0:
 
 14.0
 ----
 
-*In development*
+*November 9, 2024*
 
 Backwards-incompatible changes
 ..............................
@@ -41,7 +81,7 @@ Backwards-incompatible changes
     websockets 13.1 is the last version supporting Python 3.8.
 
 .. admonition:: The new :mod:`asyncio` implementation is now the default.
-    :class: caution
+    :class: danger
 
     The following aliases in the ``websockets`` package were switched to the new
     :mod:`asyncio` implementation::
@@ -64,8 +104,20 @@ Backwards-incompatible changes
     The :doc:`upgrade guide <../howto/upgrade>` provides complete instructions
     to migrate your application.
 
-    Aliases for deprecated API were removed from ``__all__``. As a consequence,
-    they cannot be imported e.g. with ``from websockets import *`` anymore.
+    Aliases for deprecated API were removed from ``websockets.__all__``, meaning
+    that they cannot be imported with ``from websockets import *`` anymore.
+
+.. admonition:: Several API raise :exc:`ValueError` instead of :exc:`TypeError`
+    on invalid arguments.
+    :class: note
+
+    :func:`~asyncio.client.connect`, :func:`~asyncio.client.unix_connect`, and
+    :func:`~asyncio.server.basic_auth` in the :mod:`asyncio` implementation as
+    well as :func:`~sync.client.connect`, :func:`~sync.client.unix_connect`,
+    :func:`~sync.server.serve`, :func:`~sync.server.unix_serve`, and
+    :func:`~sync.server.basic_auth` in the :mod:`threading` implementation now
+    raise :exc:`ValueError` when a required argument isn't provided or an
+    argument that is incompatible with others is provided.
 
 .. admonition:: Several API raise :exc:`ValueError` instead of :exc:`TypeError`
     for invalid arguments.
@@ -83,10 +135,16 @@ Backwards-incompatible changes
     :class: note
 
     In addition to :class:`bytes`, it may be a :class:`bytearray` or a
-    :class:`memoryview`.
+    :class:`memoryview`. If you wrote an :class:`~extensions.Extension` that
+    relies on methods not provided by these types, you must update your code.
 
-    If you wrote an :class:`extension <extensions.Extension>` that relies on
-    methods not provided by these new types, you may need to update your code.
+.. admonition:: The signature of :exc:`~exceptions.PayloadTooBig` changed.
+    :class: note
+
+    If you wrote an extension that raises :exc:`~exceptions.PayloadTooBig` in
+    :meth:`~extensions.Extension.decode`, for example, you must replace
+    ``PayloadTooBig(f"over size limit ({size} > {max_size} bytes)")`` with
+    ``PayloadTooBig(size, max_size)``.
 
 New features
 ............
@@ -94,8 +152,8 @@ New features
 * Added an option to receive text frames as :class:`bytes`, without decoding,
   in the :mod:`threading` implementation; also binary frames as :class:`str`.
 
-* Added an option to send :class:`bytes` as a text frame in the :mod:`asyncio`
-  and :mod:`threading` implementations, as well as :class:`str` a binary frame.
+* Added an option to send :class:`bytes` in a text frame in the :mod:`asyncio`
+  and :mod:`threading` implementations; also :class:`str` in a binary frame.
 
 Improvements
 ............
@@ -103,6 +161,23 @@ Improvements
 * The :mod:`threading` implementation receives messages faster.
 
 * Sending or receiving large compressed messages is now faster.
+
+* Errors when a fragmented message is too large are clearer.
+
+* Log messages at the :data:`~logging.WARNING` and :data:`~logging.INFO` levels
+  no longer include stack traces.
+
+Bug fixes
+.........
+
+* Clients no longer crash when the server rejects the opening handshake and the
+  HTTP response doesn't Include a ``Content-Length`` header.
+
+* Returning an HTTP response in ``process_request`` or ``process_response``
+  doesn't generate a log message at the :data:`~logging.ERROR` level anymore.
+
+* Connections are closed with code 1007 (invalid data) when receiving invalid
+  UTF-8 in a text frame.
 
 .. _13.1:
 
@@ -1473,7 +1548,7 @@ New features
 
 * Added support for providing and checking Origin_.
 
-.. _Origin: https://www.rfc-editor.org/rfc/rfc6455.html#section-10.2
+.. _Origin: https://datatracker.ietf.org/doc/html/rfc6455.html#section-10.2
 
 .. _2.0: